refactor(cli): move placeorder methods from utils
