+++ lib/cli/commands/buy.ts
@@ -1,0 +1,11 @@
+import { Arguments, Argv } from 'yargs';
+import { OrderSide } from '../../proto/xudrpc_pb';
+import { placeOrderBuilder, placeOrderHandler } from '../placeorder';
+
+export const command = 'buy <quantity> <pair_id> <price> [order_id] [stream]';
+
+export const describe = 'place a buy order';
+
+export const builder = (argv: Argv) => placeOrderBuilder(argv, OrderSide.BUY);
+
+export const handler = (argv: Arguments) => placeOrderHandler(argv, OrderSide.BUY);
